Split up client.ts and some small cleanups\rTESTING=unit
