+++ src/api-client.ts
@@ -1,0 +1,6 @@
+/**
+ * Simple function interface for making API calls.
+ */
+export interface ApiClient<Req, Res> {
+  (request: Req): Promise<Res>;
+}
