+++ src/metrics-request.ts
@@ -1,0 +1,21 @@
+import { InsertionMapFn } from './base-request';
+import { Insertion, Request } from './types/delivery';
+
+/**
+ * Represents a single call for logging content.
+ */
+export interface MetricsRequest {
+  /** A function to shrink the Insertions on Metrics API. */
+  toCompactMetricsInsertion?: InsertionMapFn;
+
+  /**
+   * The Request for content.
+   */
+  request: Request;
+
+  /**
+   * Insertions with all metadata.  The `toCompact` functions are used to
+   * transform the fullInsertion to Insertions on each of the requests.
+   */
+  fullInsertion: Insertion[];
+}
